requirements.txt
colorama==0.4.6
numpy==2.1.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
pandas==2.2.3
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tqdm==4.66.5
tzdata==2024.2
